requirements.txt
pyside6
pydicom
Pillow
numpy
